IGR-411 Fix casting double down to int
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@56774 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/euclidian/EuclidianView.java
@@ -3412,8 +3412,8 @@ public abstract class EuclidianView implements EuclidianViewInterfaceCommon,
 				sw = (int) Math.round(x0 * 2);
 				sh = (int) Math.round(y0 * 2);
 			}
-			int dx = (w - sw) / 2;
-			int dy = (h - sh) / 2;
+			double dx = (w - sw) / 2;
+			double dy = (h - sh) / 2;
 			xZero = getSettings().getXZero() + dx;
 			yZero = getSettings().getYZero() + dy;
 			getSettings().setSize(w, h);
